refactor(repo C14b): extract retrieval orchestrators to engine/repository/retrieval/
Move the three search orchestrators out of engine/repository.ts into their
own files under engine/repository/retrieval/:

- hybrid.ts: hybridQuery (+ SearchHooks, HybridQueryOptions,
  HybridQueryResult, StructuredSearchDiagnostics, SearchRoutingProfile)
- vector-query.ts: vectorSearchQuery (+ VectorSearchOptions,
  VectorSearchResult)
- structured.ts: structuredSearch, structuredSearchWithDiagnostics
  (+ StructuredSubSearch, StructuredSearchOptions,
  StructuredSearchWithDiagnosticsResult)

Cross-cluster dependencies (Store type, DEFAULT_EMBED_MODEL,
RERANK_CANDIDATE_LIMIT, STRONG_SIGNAL_* thresholds,
getGraphConnectedCandidates) are imported back from ../../repository.js.
Node ESM resolves these lazily, so the cycle is harmless at runtime
and Store imports remain type-only.

engine/repository.ts drops from 2040 to 952 lines. All 973 tests pass.

Spec: docs/superpowers/specs/2026-05-20-kindx-strategic-refactor-program-design.md §5
